+++ mobile/index.html
@@ -1,3 +1,3 @@
 <script type="text/javascript">
-	location.href="mobile.php?largura="+screen.width;
+	location.href="mobile.php?largura=" + screen.width;
 </script>
