+++ mobile/scripts/main.js
@@ -1,1 +1,1 @@
-!function(a){"use strict";var b={init:function(){this.header=a.document.getElementById("header"),this.all=a.document.getElementById("all"),this.events()},events:function(){var b=this;$(a).ready(function(){b.doTop(b.header,b.all)}),$(a).on("resize",function(){b.doTop(b.header,b.all)}).on("orientationchange",function(){b.doTop(b.header,b.all)})},doTop:function(a,b){var c=a.clientHeight;b.style.marginTop=c+"px",a.style.position="fixed"}};b.init()}(window),function(a,b){"use strict";var c={init:function(){this.events()},events:function(){function c(){l||(k=f.height(),f.css({height:"0"}),l=!0),e.hasClass("active")?(e.removeClass("active"),f.stop().animate({height:0},300,function(){f.hide()})):(e.addClass("active"),f.show().stop().animate({height:k},300))}function d(){b.pageYOffset>70?j.hasClass("reduce")||j.addClass("reduce"):j.removeClass("reduce")}var e=a(".js-btn-menu"),f=a(".js-menu"),g=a(".js-nav-menu"),h=a(".js-outside"),i=a(".js-btn-back-info"),j=a(".js-header"),k=0,l=!1;a(b).ready(function(){a.scrollIt({upKey:38,downKey:40,easing:"linear",scrollTime:600,activeClass:"active",onPageChange:null,topOffset:-47})}),e.on("click",function(a){c(),a.preventDefault()}),h.on("click",function(a){c(),a.preventDefault()}),g.on("click",function(a){c(),i.trigger("click"),a.preventDefault()}),b.onscroll=d}};c.init()}(jQuery,window),function(){"use strict";var a={init:function(){this.events()},events:function(){}};a.init()}(window),function(){"use strict";var a={init:function(){this.events()},events:function(){function a(a){f.unslick(),f.html(""),d.html(a.titulo),e.html(a.texto);var b="<li>";b+='<div class="info-case js-info-case">',b+='<div class="wrap-content js-wrap-content">',b+='<div class="content js-content">',b+='<div class="fix-content">',b+='<h3 class="js-title">'+a.titulo+"</h3>",b+='<div class="js-text">'+a.texto+"</div>",b+="</div>",b+="</div>",b+='<h3 class="js-title"></h3>',b+='<div class="js-text"></div>',b+="</div>",b+="</div>",b+="</li>",f.append(b),a.assets.forEach(function(a){var b="<li>";if(a.url.length>0){var c=a.url.indexOf("v=");b+='<div class="thumb">',b+='<div class="wrap-video">',b+='<iframe class="js-video" src="//www.youtube.com/embed/'+a.url.slice(c+2)+'?rel=0&amp;controls=0" width="960" height="540" frameborder="0" allowfullscreen></iframe>',b+="</div>",b+="</div>"}else{var d=0;d=screenWidth>=0&&640>=screenWidth?a.img_640:screenWidth>=640&&1024>=screenWidth?a.img_1024:a.img,b+='<figure class="thumb">',b+='<img class="img" src="'+d+'" alt="">',b+="</figure>"}b+="</li>",f.append(b)}),f.slick({dots:!0,infinite:!0,speed:300,lazyLoad:!1,slidesToShow:1,slidesToScroll:1,slide:"li",swipe:!0,prevArrows:'<button type="button" class="slick-prev arrow-left">Previous</button>',nextArrows:'<button type="button" class="slick-next arrow-right">Next</button>'}),$(".js-box-video").on("click",function(){})}var b=$(".js-thumb-case"),c=$(".js-gallery"),d=$(".js-title"),e=$(".js-text"),f=$(".js-list-cases"),g=$(".js-new");b.on("click",function(){var b=$(this).data("category-id");for(var d in Cases)Cases.hasOwnProperty(d)&&b==parseInt(Cases[d].id)&&a(Cases[d]);c.stop().animate({left:"-100%"},300)}),g.on("click",function(){var b=$(this).data("category-id");for(var d in Cases)Cases.hasOwnProperty(d)&&b==parseInt(Cases[d].id)&&a(Cases[d]);c.stop().animate({left:"-100%"},300)})}};a.init()}(window),function(){"use strict";var a={init:function(){this.events()},events:function(){{var a=$(".js-case"),b=a.find(".js-back"),c=$(".js-gallery");$(".js-info-case"),$(".js-wrap-content"),$(".js-btn-info"),$(".js-btn-back-info"),$(".js-wrap-case"),$(".fix-content")}b.on("click",function(){c.stop().animate({left:"0"},300)})}};a.init()}(window);var Video=function(){"use strict";var a={init:function(a){this.events(a)},events:function(a){function b(a){$(a.currentTarget).siblings("*")[0].paused?($(a.currentTarget).find(".js-btn-play").fadeOut(300),$(a.currentTarget).siblings("*")[0].play()):$(a.currentTarget).siblings("*")[0].paused||($(a.currentTarget).find(".js-btn-play").show(),$(a.currentTarget).siblings("*")[0].pause())}var c=a,d=c.siblings(".js-btn-video"),e=d.find(".js-btn-play");d.on("click",b),d.on("touchstart",b),c.on("ended",function(){e.fadeIn(300)})}};return a}(window);Video.init($(".js-video")),function(a,b){"use strict";var c={init:function(){this.events()},events:function(){function c(b){var c=a("#content-"+b.attr("id"));c.siblings("*").stop().animate({height:0},300),b.closest(".js-type").siblings("*").find(".active").removeClass("active"),b.hasClass("active")?(b.removeClass("active"),c.stop().animate({height:0},300)):(b.addClass("active"),c.stop().animate({height:c.data("height")},300))}function d(){e.each(function(){var b=a(this),c=a("#content-"+b.attr("id"));b.hasClass("active")?c.data("height",c.height()).css({height:"auto"}):c.css({display:"none",height:"auto"}).data("height",c.height()).css({height:"0",display:"block"})})}var e=a(".js-btn-contact");e.each(function(){var b=a(this),d=a("#content-"+b.attr("id"));b.on("click",function(b){c(a(this)),b.preventDefault()}),d.data("height",d.height()).css({height:"0",display:"block"})}),b.addEventListener("resize",d)}};c.init()}(jQuery,window),function(){"use strict";var a={init:function(){this.events()},events:function(){function a(a){NProgress.done();var c='<div class="icon-wran success"></div>';c+='<span class="text-warn">sua mensagem <br> foi enviada</span>';var d='<div class="icon-wran error"></div>';d+='<span class="text-warn">OPS! <br > Sua mensagem <br >não foi enviada. <br >Toque para <br > tentar novamente.</span>',k.css({visibility:"hidden"}),a?(l.html(c),l.off("click"),l.on("click",b)):(l.html(d),l.off("click"),l.on("click",function(){m||g()}))}function b(){l.html("");var a=i.find('[name="nome"]'),b=i.find('[name="email"]'),c=i.find('[name="descricao"]');a.val(""),b.val(""),c.val(""),k.css({visibility:"visible"})}function c(a){var b=i.find('[name="nome"]'),c=i.find('[name="email"]'),h=i.find('[name="descricao"]');e(b,"Menor que 3 caracteres."),f(c,"E-mail inválido."),d(h,"Campo vazio."),0===$(".invalid").length&&a&&g()}function d(a,b){var c=/([^\s])/,d=c.test(a.val());return h(a,b,d),d}function e(a,b){var c=/([^\s])/,d=c.test(a.val()),e=a.val().length>2?!0:!1,f=!1;return f=d&&e?!0:!1,h(a,b,f),f}function f(a,b){var c=/^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/,d=c.test(a.val());return h(a,b,d),d}function g(){m=!0,j.addClass("sending"),NProgress.start(),$.ajax({type:"POST",url:"/service/contatos",data:i.serialize(),dataType:"json",success:function(b){a(200===parseInt(b.cod)?!0:!1)},error:function(){a(!1)}}).done(function(){m=!1,j.removeClass("sending"),$(".js-header").removeClass("fixed-header")})}function h(a,b,c){var d=a.siblings(".js-warn"),e=d.find(".js-warn-text");c?(a.removeClass("invalid"),d.removeClass("error"),e.html("")):(a.addClass("invalid"),d.addClass("error"),e.html(b))}var i=$(".js-form-contact"),j=i.find(".js-submit"),k=i.find(".js-wrap-content-form"),l=i.find(".js-wrap-response-status"),m=!1;j.on("click",function(a){i.addClass("catch"),m||c(!0),a.preventDefault()}),i.find("input, textarea").mouseover(function(){$(".js-header").addClass("fixed-header")}),i.on("keyup","input, textarea",function(a){$(".catch").length>0&&c(!1),a.preventDefault()}).on("blur","input, textarea",function(a){$(".js-header").removeClass("fixed-header"),$(".catch").length>0&&c(!1),a.preventDefault()})}};a.init()}(window);
+!function(a){"use strict";var b={init:function(){this.header=a.document.getElementById("header"),this.all=a.document.getElementById("all"),this.events()},events:function(){var b=this;$(a).ready(function(){b.doTop(b.header,b.all)}),$(a).on("resize",function(){b.doTop(b.header,b.all)}).on("orientationchange",function(){b.doTop(b.header,b.all)})},doTop:function(a,b){var c=a.clientHeight;b.style.marginTop=c+"px",a.style.position="fixed"}};b.init()}(window),function(a,b){"use strict";var c={init:function(){this.events()},events:function(){function c(){l||(k=f.height(),f.css({height:"0"}),l=!0),e.hasClass("active")?(e.removeClass("active"),f.stop().animate({height:0},300,function(){f.hide()})):(e.addClass("active"),f.show().stop().animate({height:k},300))}function d(){b.pageYOffset>70?j.hasClass("reduce")||j.addClass("reduce"):j.removeClass("reduce")}var e=a(".js-btn-menu"),f=a(".js-menu"),g=a(".js-nav-menu"),h=a(".js-outside"),i=a(".js-btn-back-info"),j=a(".js-header"),k=0,l=!1;a(b).ready(function(){a.scrollIt({upKey:38,downKey:40,easing:"linear",scrollTime:600,activeClass:"active",onPageChange:null,topOffset:-47})}),e.on("click",function(a){c(),a.preventDefault()}),h.on("click",function(a){c(),a.preventDefault()}),g.on("click",function(a){c(),i.trigger("click"),a.preventDefault()}),b.onscroll=d}};c.init()}(jQuery,window),function(){"use strict";var a={init:function(){this.events()},events:function(){}};a.init()}(window),function(){"use strict";var a={init:function(){this.events()},events:function(){function a(a){f.unslick(),f.html(""),d.html(a.titulo),e.html(a.texto);var b="<li>";b+='<div class="info-case js-info-case">',b+='<div class="wrap-content js-wrap-content">',b+='<div class="content js-content">',b+='<div class="fix-content">',b+='<h3 class="js-title">'+a.titulo+"</h3>",b+='<div class="js-text">'+a.texto+"</div>",b+="</div>",b+="</div>",b+='<h3 class="js-title"></h3>',b+='<div class="js-text"></div>',b+="</div>",b+="</div>",b+="</li>",f.append(b),a.assets.forEach(function(a){var b="<li>";if(a.url.length>0){var c=a.url.indexOf("v=");b+='<div class="thumb">',b+='<div class="wrap-video">',b+='<iframe class="js-video" src="//www.youtube.com/embed/'+a.url.slice(c+2)+'?rel=0&amp;controls=0" width="960" height="540" frameborder="0" allowfullscreen></iframe>',b+="</div>",b+="</div>"}else{var d=0;d=screenWidth>=0&&640>=screenWidth?a.img_640:screenWidth>=640&&1024>=screenWidth?a.img_1024:a.img,b+='<figure class="thumb">',b+='<img class="img" src="'+d+'" alt="">',b+="</figure>"}b+="</li>",f.append(b)}),f.slick({dots:!0,infinite:!0,speed:300,lazyLoad:!1,slidesToShow:1,slidesToScroll:1,slide:"li",swipe:!0,prevArrows:'<button type="button" class="slick-prev arrow-left">Previous</button>',nextArrows:'<button type="button" class="slick-next arrow-right">Next</button>'}),$(".js-box-video").on("click",function(){})}var b=$(".js-thumb-case"),c=$(".js-gallery"),d=$(".js-title"),e=$(".js-text"),f=$(".js-list-cases"),g=$(".js-new");b.on("click",function(){var b=$(this).data("category-id");for(var d in Cases)Cases.hasOwnProperty(d)&&b==parseInt(Cases[d].id)&&a(Cases[d]);c.stop().animate({left:"-100%"},300)}),g.on("click",function(){var b=$(this).data("category-id");for(var d in Cases)Cases.hasOwnProperty(d)&&b==parseInt(Cases[d].id)&&a(Cases[d]);c.stop().animate({left:"-100%"},300)})}};a.init()}(window),function(){"use strict";var a={init:function(){this.events()},events:function(){{var a=$(".js-case"),b=a.find(".js-back"),c=$(".js-gallery");$(".js-info-case"),$(".js-wrap-content"),$(".js-btn-info"),$(".js-btn-back-info"),$(".js-wrap-case"),$(".fix-content")}b.on("click",function(){c.stop().animate({left:"0"},300)})}};a.init()}(window);var Video=function(){"use strict";var a={init:function(a){this.events(a)},events:function(a){function b(a){$(a.currentTarget).siblings("*")[0].paused?($(a.currentTarget).find(".js-btn-play").fadeOut(300),$(a.currentTarget).siblings("*")[0].play()):$(a.currentTarget).siblings("*")[0].paused||($(a.currentTarget).find(".js-btn-play").show(),$(a.currentTarget).siblings("*")[0].pause())}var c=a,d=c.siblings(".js-btn-video"),e=d.find(".js-btn-play");d.on("click",b),d.on("touchstart",b),c.on("ended",function(){e.fadeIn(300)})}};return a}(window);Video.init($(".js-video")),function(a,b){"use strict";var c={init:function(){this.events()},events:function(){function c(b){var c=a("#content-"+b.attr("id"));c.siblings("*").stop().animate({height:0},300),b.closest(".js-type").siblings("*").find(".active").removeClass("active"),b.hasClass("active")?(b.removeClass("active"),c.stop().animate({height:0},300)):(b.addClass("active"),c.stop().animate({height:c.data("height")},300))}function d(){e.each(function(){var b=a(this),c=a("#content-"+b.attr("id"));b.hasClass("active")?c.data("height",c.height()).css({height:"auto"}):c.css({display:"none",height:"auto"}).data("height",c.height()).css({height:"0",display:"block"})})}var e=a(".js-btn-contact");e.each(function(){var b=a(this),d=a("#content-"+b.attr("id"));b.on("click",function(b){c(a(this)),b.preventDefault()}),d.data("height",d.height()).css({height:"0",display:"block"})}),b.addEventListener("resize",d)}};c.init()}(jQuery,window),function(){"use strict";var a={init:function(){this.events()},events:function(){function a(a){NProgress.done();var c='<div class="icon-wran success"></div>';c+='<span class="text-warn">sua mensagem <br> foi enviada</span>';var d='<div class="icon-wran error"></div>';d+='<span class="text-warn">OPS! <br > Sua mensagem <br >não foi enviada. <br >Toque para <br > tentar novamente.</span>',k.css({visibility:"hidden"}),a?(l.html(c),l.off("click"),l.on("click",b)):(l.html(d),l.off("click"),l.on("click",function(){m||g()}))}function b(){l.html("");var a=i.find('[name="nome"]'),b=i.find('[name="email"]'),c=i.find('[name="descricao"]');a.val(""),b.val(""),c.val(""),k.css({visibility:"visible"})}function c(a){var b=i.find('[name="nome"]'),c=i.find('[name="email"]'),h=i.find('[name="descricao"]');e(b,(LANG == 'pt' ? "Menor que 3 caracteres." : "Field invalid.")),f(c,(LANG == 'pt' ? "E-mail inválido." : "E-mail is invalid.")),d(h,(LANG == 'pt' ? "Campo vazio." : "Field is empty.")),0===$(".invalid").length&&a&&g()}function d(a,b){var c=/([^\s])/,d=c.test(a.val());return h(a,b,d),d}function e(a,b){var c=/([^\s])/,d=c.test(a.val()),e=a.val().length>2?!0:!1,f=!1;return f=d&&e?!0:!1,h(a,b,f),f}function f(a,b){var c=/^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/,d=c.test(a.val());return h(a,b,d),d}function g(){m=!0,j.addClass("sending"),NProgress.start(),$.ajax({type:"POST",url:"/service/contatos",data:i.serialize(),dataType:"json",success:function(b){a(200===parseInt(b.cod)?!0:!1)},error:function(){a(!1)}}).done(function(){m=!1,j.removeClass("sending"),$(".js-header").removeClass("fixed-header")})}function h(a,b,c){var d=a.siblings(".js-warn"),e=d.find(".js-warn-text");c?(a.removeClass("invalid"),d.removeClass("error"),e.html("")):(a.addClass("invalid"),d.addClass("error"),e.html(b))}var i=$(".js-form-contact"),j=i.find(".js-submit"),k=i.find(".js-wrap-content-form"),l=i.find(".js-wrap-response-status"),m=!1;j.on("click",function(a){i.addClass("catch"),m||c(!0),a.preventDefault()}),i.find("input, textarea").mouseover(function(){$(".js-header").addClass("fixed-header")}),i.on("keyup","input, textarea",function(a){$(".catch").length>0&&c(!1),a.preventDefault()}).on("blur","input, textarea",function(a){$(".js-header").removeClass("fixed-header"),$(".catch").length>0&&c(!1),a.preventDefault()})}};a.init()}(window);
