+++ mobile/styles/main.css
@@ -598,7 +598,7 @@ body {
 }
 #news img {
     height:100%;
-    margin-left:-36%
+    margin-left:-24%
 }
 #gallery {
     position:relative;
